+++ src/main/java/pl/marcool/intivepatronage/services/OrganizationService.java
@@ -4,41 +4,45 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.marcool.intivepatronage.models.Organization;
 import pl.marcool.intivepatronage.repositores.OrganizationRepository;
-import java.util.List;
+
+import java.util.Optional;
 
 @Service
 public class OrganizationService {
 
     @Autowired
     private OrganizationRepository organizationRepository;
-    private String message = "The following error was detected:\n";
 
     public void save(Organization organization) throws MyExceptions {
-        String isIdExist = organizationRepository.findById(organization.getName()).getName();
-        if (!isIdExist.equals("empty"))
-            throw new MyExceptions(message + "Organization '" + isIdExist + "' already exists");
+        if (organizationRepository.findById(organization.getName()).isPresent())
+            throw new MyExceptions("Organization ID:" + organization.getName() + " - already exists.");
         organizationRepository.save(organization);
     }
 
-    public List getAll() {
+    public Iterable<Organization> getAll() {
         return organizationRepository.findAll();
     }
 
-    public Organization findById(String id) {
-        return organizationRepository.findById(id);
+    public Organization findById(String id) throws MyExceptions {
+        Optional<Organization> organization = organizationRepository.findById(id);
+        if (organization.isEmpty()) throw new MyExceptions("Organization ID:" + id + " - not found.");
+        return organization.get();
     }
 
     public void update(String id, Organization organization) throws MyExceptions {
-        String isOldNameExist = findById(id).getName();
-        String isNewNameExist = findById(organization.getName()).getName();
-        if (isOldNameExist.equals("empty"))
-            throw new MyExceptions(message + "No organization '" + id + "' was found");
-        if (!isNewNameExist.equals("empty") & !isNewNameExist.equals(isOldNameExist))
-            throw new MyExceptions(message + "Name '" + isNewNameExist + "' already exists");
-        organizationRepository.update(id, organization);
+        if (organizationRepository.findById(id).isEmpty())
+            throw new MyExceptions("Organization ID:" + id + " - not found.");
+        if (organizationRepository.findById(organization.getName())
+                .stream()
+                .anyMatch(cr -> !cr.getName().equals(id)))
+            throw new MyExceptions("Organization ID:" + organization.getName() + " - already exists.");
+        organizationRepository.deleteById(id);
+        organizationRepository.save(organization);
     }
 
-    public void deleteById(String id) {
+    public void deleteById(String id) throws MyExceptions {
+        if (organizationRepository.findById(id).isEmpty())
+            throw new MyExceptions("Organization ID:" + id + " - not found.");
         organizationRepository.deleteById(id);
     }
